Add streaming C implementation
Port the streaming version of dedup.py to C.  This is ~50x faster: the 1hr runtime for 2M reads I was observing with the streaming version is now under two minutes.

For tests I decided to run all tests on both the python and C version, to avoid duplicate tests and ensure we keep them in sync.

Written by Claude Code, with lots of prompting and some manual fixups.