Add streaming Rust implementation (#5)
Port the streaming API of dedup.py to Rust. This is ~50x faster: the 1hr runtime for 2M reads I was observing with the streaming version is now just over one minute.

I originally did this in C (https://github.com/securebio/nao-dedup/pull/4) but after discussion with @evanfields and @willbradshaw we decided Rust was a better choice.  I'm much less familiar with Rust, so this relies very heavily on Claude Code: I'm not just using to to quickly write something I could write myself, but instead to translate my C and Python into a language I don't really know.

All streaming tests run against both python and rust, to avoid duplicate tests and ensure we keep implementations in sync.

I think it's good to keep both the python and rust versions of the streaming code, because organizationally we're not nearly as strong in rust and rust is less readable, so the python version does double duty as a kind of very detailed documentation of the approach.